+++ changelog.md
@@ -1,11 +1,19 @@
-## 0.1.1
+#todo
+
+#dev-master
+
+ - added methods to make URL query handling easier: `WebUtil::removeQuery()` and `WebUtil::getQuery()`
+ - updated to newest version of [glenscott/url-normalizer](https://github.com/glenscott/url-normalizer/) once dev-master gets tagged due to [issue #17: SplFileObject Flags have no effect / empty lines cannot be ignored](https://github.com/glenscott/url-normalizer/issues/17)
+ - updated to Guzzle ~5.3
+ 
+##0.1.1
 
  - updated repositories to local satis installation
 
-## 0.1.0
+##0.1.0
 
  - changed package name from WebUtility to web-utility
 
 ###0.0.1
 
-- Initial commit
+ - initial commit